requirements.txt
SpeechRecognition
llama-cpp-python
pyttsx3